homework6/pyproject.toml
[tool.poetry]
name = "homework6"
version = "0.1.0"
description = ""
authors = ["Aryna Reutska <arina.reutskaya110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
